# docs/swagger.yaml
basePath: /
definitions:
  httpapi.APIError:
    properties:
      error:
        properties:
          code:
            example: VALIDATION_ERROR
            type: string
          message:
            example: movement input is invalid
            type: string
        type: object
    type: object
  httpapi.DataResponse:
    properties:
      data: {}
    type: object
  httpapi.ListResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/operations.Page'
    type: object
  httpapi.movementRequest:
    properties:
      direction:
        allOf:
        - $ref: '#/definitions/movement.Direction'
        description: Required for all movements. IN requires inCategory; OUT requires
          outDestination and its matching fields.
        enum:
        - IN
        - OUT
        example: OUT
      driverId:
        description: Required only when outDestination is OTHER.
        example: 42
        type: integer
      driverName:
        description: Required only when outDestination is OTHER.
        example: Driver Name
        type: string
      inCategory:
        description: Required only for IN. Must not be supplied for OUT.
        enum:
        - WAITING_FOR_ASSIGNMENT
        - WAITING_FOR_DELIVERY_TO_CUSTOMER
        - SERVICE_AND_REPAIR_MAINTENANCE
        type: string
      noLambung:
        description: Required for all movements.
        example: SAMT223
        type: string
      note:
        description: Required only when outDestination is OTHER; optional for other
          OUT destinations.
        example: Operational note
        type: string
      outDestination:
        description: Required only for OUT. Must not be supplied for IN.
        enum:
        - BENGKEL_LUAR
        - FILLING_SHED_KUIN
        - CUSTOMER
        - OTHER
        type: string
      sppNumber:
        description: Required only when outDestination is FILLING_SHED_KUIN or CUSTOMER.
        example: SPP-001
        type: string
      workOrderNumber:
        description: Required only when outDestination is BENGKEL_LUAR.
        example: WO-001
        type: string
    required:
    - direction
    - noLambung
    type: object
...
  movement.Transaction:
    properties:
      actorId:
        type: string
      actorName:
        type: string
      customerId:
        type: string
        x-nullable: "true"
      customerName:
        type: string
        x-nullable: "true"
      direction:
        $ref: '#/definitions/movement.Direction'
      driverId:
        type: integer
        x-nullable: "true"
      driverName:
        type: string
        x-nullable: "true"
      inCategory:
        type: string
        x-nullable: "true"
      lambungUnitId:
        type: integer
      noLambung:
        type: string
      noPolisi:
        type: string
      note:
        type: string
        x-nullable: "true"
      occurredAt:
        type: string
      outDestination:
        type: string
        x-nullable: "true"
      sppNumber:
        type: string
        x-nullable: "true"
      transactionId:
        type: integer
      workOrderNumber:
        type: string
        x-nullable: "true"
    type: object
  operations.Dashboard:
    properties:
      totalUnits:
        type: integer
      unitsIn:
        type: integer
      unitsNoStatus:
        type: integer
      unitsOut:
        type: integer
    type: object
...
  operations.Transaction:
    properties:
      actorName:
        type: string
      customerName:
        type: string
        x-nullable: "true"
      direction:
        type: string
      driverName:
        type: string
        x-nullable: "true"
      inCategory:
        type: string
        x-nullable: "true"
      noLambung:
        type: string
      noPolisi:
        type: string
      note:
        type: string
        x-nullable: "true"
      occurredAt:
        type: string
      outDestination:
        type: string
        x-nullable: "true"
      sppNumber:
        type: string
        x-nullable: "true"
      transactionId:
        type: integer
      workOrderNumber:
        type: string
        x-nullable: "true"
    type: object
...
info:
  contact: {}
  description: |-
    API for immutable truck IN/OUT movements and pool operations.
    Movement scans require `unitlog:app:default:view`, actor ID/name claims, and an Idempotency-Key UUID. Operations require `unitlog:web:default:view`.
  title: UnitLog Pool Basirih API
  version: 1.0.0
paths:
  /api/v1/dashboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/operations.Dashboard'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: Get operational dashboard totals
      tags:
      - Operations
  /api/v1/movements:
    post:
      consumes:
      - application/json
      description: Records an IN or OUT transaction. Retries must reuse the same Idempotency-Key
        and request body. IN requires inCategory only. OUT requires BENGKEL_LUAR with
        workOrderNumber, FILLING_SHED_KUIN or CUSTOMER with sppNumber, or OTHER with
        driverId, driverName, and note.
      parameters:
      - description: UUID generated once per movement
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Movement request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.movementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/movement.Transaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: Record a unit movement
      tags:
      - Movements
  /api/v1/transactions:
    get:
      parameters:
      - description: Search no_lambung or no_polisi
        in: query
        name: query
        type: string
      - description: Movement direction
        enum:
        - IN
        - OUT
        in: query
        name: direction
        type: string
      - description: IN category or OUT destination
        in: query
        name: detail
        type: string
      - description: Inclusive RFC3339 UTC timestamp
        in: query
        name: from
        type: string
      - description: Exclusive RFC3339 UTC timestamp
        in: query
        name: to
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/operations.Transaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - Operations
  /api/v1/transactions/export:
    get:
      parameters:
      - description: Search no_lambung or no_polisi
        in: query
        name: query
        type: string
      - description: Movement direction
        enum:
        - IN
        - OUT
        in: query
        name: direction
        type: string
      - description: IN category or OUT destination
        in: query
        name: detail
        type: string
      - description: Inclusive RFC3339 UTC timestamp
        in: query
        name: from
        type: string
      - description: Exclusive RFC3339 UTC timestamp
        in: query
        name: to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: Export transactions
      tags:
      - Operations
  /api/v1/units:
    get:
      parameters:
      - description: Search no_lambung or no_polisi
        in: query
        name: query
        type: string
      - description: Current direction
        enum:
        - IN
        - OUT
        - "NULL"
        in: query
        name: status
        type: string
      - description: Current OUT destination
        enum:
        - BENGKEL_LUAR
        - FILLING_SHED_KUIN
        - CUSTOMER
        - OTHER
        in: query
        name: destination
        type: string
      - description: Include inactive units
        in: query
        name: includeInactive
        type: boolean
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/operations.Unit'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: List units
      tags:
      - Operations
  /api/v1/units/{id}:
    get:
      parameters:
      - description: Unit Lambung ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/operations.Unit'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: Get unit detail
      tags:
      - Operations
  /api/v1/units/{id}/transactions:
    get:
      parameters:
      - description: Unit Lambung ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 25
        description: Page size
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpapi.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/operations.Transaction'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: List unit transactions
      tags:
      - Operations
  /api/v1/units/export:
    get:
      parameters:
      - description: Search no_lambung or no_polisi
        in: query
        name: query
        type: string
      - description: Current direction
        enum:
        - IN
        - OUT
        - "NULL"
        in: query
        name: status
        type: string
      - description: Current OUT destination
        enum:
        - BENGKEL_LUAR
        - FILLING_SHED_KUIN
        - CUSTOMER
        - OTHER
        in: query
        name: destination
        type: string
      - description: Include inactive units
        in: query
        name: includeInactive
        type: boolean
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.APIError'
      security:
      - BearerAuth: []
      summary: Export units
      tags:
      - Operations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Auth Service JWT using the `Bearer <token>` format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"